workqueue: add delayed metric
Reporting the number of items currently waiting for retry. This metric
is suitable for an alert if it is above 0 for a long time.

Previous retry metrics is not suitable for alerting because it is hard
to distinguish a random transit error from a backoff repeating error.

Implemented as new DelayedMetricsProvider interface to avoid breaking
existing external provider.