client-go cache: avoid API break in Controller interface
Adding methods to an exported interface is an API break because it would force
out-of-tree implementers to adapt their implementation. It's unclear whether
those exist, but it's not hard to play it safe and avoid extending the
interface.

Because HasSynced was public and might be used, NamedController gets added to
also expose NamedHasSynced.