pkg/scheduler: remove duplicated "pod" key in scheduler logging
Now that the pod is added to the context with klog.LoggerWithValues
at the beginning of the scheduling cycle, passing the redundant
"pod", klog.KObj(pod) key-value pair in every log line is no longer
necessary.

This change removes all remaining duplicated "pod" keys
in pkg/scheduler to fully migrate to contextual logging.

This is part of the ongoing effort to adopt contextual logging in the
scheduler (KEP-3071).