Simplify nftables masquerade rule
iptables kube-proxy's masquerade rule is:

  -A KUBE-POSTROUTING -m mark ! --mark 0x4000/0x4000 -j RETURN
  -A KUBE-POSTROUTING -j MARK --xor-mark 0x4000
  -A KUBE-POSTROUTING -j MASQUERADE

That is "if the mark bit is not set, return from KUBE-POSTROUTING,
otherwise clear the mark bit and masquerade". The nftables backend
copied this:

  add rule ip kube-proxy masquerading mark and 0x4000 == 0 return
  add rule ip kube-proxy masquerading mark set mark xor 0x4000
  add rule ip kube-proxy masquerading masquerade fully-random

However, unsetting the mark is only needed to work around a Linux
kernel bug that was fixed in 4.19, which is much older than the oldest
kernel the nftables backend supports. So remove that and simplify the
rule.