client-go: remove GetController + Controller interface
The interface serves no useful purpose. There is not API in client-go which
accepts it as parameter, so there is no need to support out-of-tree
implementations of it. A search showed that the only out-of-tree usage seem to
be forks of the Kubernetes code.

The minimize the impact in code which merely passes it around,
`cache.Controller` continues to be valid. It's now a type alias for a pointer
to the implementation instead of an interface.

The in-tree usage of it was to have access to `HasSynced` and
`LastSyncResourceVersion`. This gets replaced with the more explicit
`cache.DoneChecker` and a function handle.

GetController already was deprecated and declared useless ages ago. This
follows through with removing it.