WIP: client-go informers: WaitForCacheSync now also waits for event handlers
This is a common mistake:

    /* handle, err := */ informerFactory....AddEventHandler(...)
    informerFactory.Start(...)
    informerFactory.WaitForCacheSync(...)
    // Do something with data populated by event handlers.

Contrary to what developers expect, WaitForCacheSync returns *before*
all initial Add events are processed by the event handler.

What developers need to do is an explicit:

    cache.WaitForCacheSync(handle.HasSynced)

This happened to work in the past when informerFactory.WaitForCacheSync as
based on polling, because it often slept at least once, giving event handlers
time to catch up.

Now that waiting is implemented through channels, WaitForCacheSync returns much
faster and tests with the mistake above might start to fail more often (remains
to be seen!).

Instead of identifying all broken tests and code (an exercise that all
downstream consumers of client-go would need to repeat!), the behavior of
informerFactory.WaitForCacheSync is now what developers have expected: it waits
for the informer caches and event handlers to be synced.

This only works for event handlers that are registered at the time when
informerFactory.WaitForCacheSync is called. When adding event handlers later,
the caller still needs to do a manual cache.WaitForCacheSync.