make test: run compilation under nice
Running tool commands (compilation, linking) under nice when available
reduces the risk that those commands steal CPU time from test binaries
running in parallel. Test binaries are sometimes senstive to real-world
time and flake when blocked from running for too long.

When used on a desktop system it keeps the desktop environment more
responsive.

This is enabled unconditionally when possible (= nice command available)
because it shouldn't be necessary to disable or configure it and less
option in the script is simpler.